+++ CHANGELOG.md
@@ -1,9 +1,22 @@
+# 0.10.72 Release:
+## New Content:
+- Banishment
+- Bless
+- Wrathful Smite (Thanks Michael)
+## Bug Fixes:
+- Blink
+- Hold Person
+- Scorching Ray
+- Shocking Grasp
+- Hunter's Mark
+## Update Notes:
+- Requires D&D system version 3.1.2 or greater. Core concentration and damage application is not supported yet.
+- Support for core concentration and damage application will be in a future release.
+  
 # 0.10.71 Release:
 ## Update Notes:
 - Updated Midi-QOl overtime creator to support new action save types.
 - Corrected Chaos Bolt DDBI import options.
-- Requires D&D system version 3.1.2 or greater. Core concentration and damage application is not supported yet.
-- Support for core concentration and damage application will be in a future release.
   
 # 0.10.70 Release:
 ## New Monster Content: